Fixed RSpec setup by simplifying the helpers
This is initial effort at simplifying the RSpec setup so specs can be added for building out the test suite. I also removed the sample spec since it was throwing errors and wasn't testing anything.

The Rails helper still needs work, though, which will need to be addressed in future commits.

Milestone: patch